pyproject.toml
[tool.poetry]
name = "appsec"
version = "0.1.0"
description = ""
authors = ["alllexxx1 <alekkorpachyov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
